synq: fold source range into FieldValue::Span and split parser.c
Follow-up cleanups on top of PR #86 ("rework macro expansion span tracking").

**Source ranges on FieldValue::Span.**  `extract_field_value` already calls
C's `resolve_span` which returns the source offset + length alongside the
text.  Before this change we threw those two fields away and forced every
consumer to call a second `field_source_range(node, idx)` method — a
redundant C roundtrip AND a verbose two-step dance at every call site.

Now `FieldValue::Span` carries a `source: SourceRange { start, end }` that
is populated in the same `resolve_span` call.  The public
`AnyParsedStatement::field_source_range` method (and the internal
`field_span` helper it relied on) is gone; `analyzer.rs`, `catalog.rs`,
`lineage/resolver.rs`, and `fmt/interpret.rs` destructure the range
directly from the pattern match.  The formatter's comment-drain logic in
particular no longer does `ptr - source.as_ptr()` pointer arithmetic,
which was latently broken for expansion-buffer spans.

**`SyntaqliteSourceSpan` fast-path helper.**  Raw spans are safe to use
directly when the input has no macro expansions.  Added
`synq_span_needs_resolve()` as an inline in `types.h` so C consumers can
branch on the fast path, and taught `syntaqlite_parser_resolve_span` to
use it — direct spans now skip the `resolve_to_source` walk entirely.

**parser.c split.**  The 1673-line `parser.c` was getting hard to read
since PR #86 landed the macro expansion pipeline.  Split into:

- `parser.c` (669 lines) — lifecycle, main parse loop, result accessors,
  incremental feed_token API, config, arena accessors
- `parser_macros.c` (760 lines) — macro registry CRUD, template
  expansion, scan_macro_args, try_macro_call, straddle diagnostic,
  begin/end_macro, resolve_span, expansion_traceback
- `parser_dump.c` (177 lines) — self-contained AST dump
- `parser_internal.h` (186 lines) — shared struct definition, macro
  types, cross-file helper declarations

Types defined only internally are consistently `Synq`-prefixed now —
`SyntaqliteMacroEntry → SynqMacroEntry`, `MacroExpansion →
SynqMacroExpansion`.  The public C ABI is unchanged.

**Doc cleanups.**  Updated three stale references to
`syntaqlite_parser_resolve_span_for_node` (renamed during PR #86).
Rewrote the `SyntaqliteSourceSpan` comment in types.h to accurately
describe the fast path instead of marking the struct opaque.
