+++ syntaqlite/src/fmt/interpret.rs
@@ -213,20 +213,25 @@ impl Formatter {
                 FmtOp::Span(idx) => {
                     // INVARIANT: Span ops only target Span fields.
                     let FieldValue::Span {
-                        text: s, quoted, ..
+                        text: s,
+                        quoted,
+                        source: span_src,
                     } = fields[idx as usize]
                     else {
                         panic!("Span: field {idx} is not a Span");
                     };
 
                     if !s.is_empty() {
-                        let offset = byte_offset_in(source, s.as_ptr());
                         if let Some(ref cctx) = ctx.comment_ctx {
                             // For quoted spans, the original token in source
                             // starts one byte earlier (the opening quote) and
                             // extends one byte past (the closing quote).
-                            let drain_offset = if quoted { offset - 1 } else { offset };
-                            let token_len = usize_to_u32(s.len()) + if quoted { 2 } else { 0 };
+                            let drain_offset = if quoted {
+                                span_src.start - 1
+                            } else {
+                                span_src.start
+                            };
+                            let token_len = span_src.len() + if quoted { 2 } else { 0 };
                             let drain = cctx.drain_before(drain_offset, source, arena);
                             flush_drain(&drain, &mut pending, &mut running, arena);
                             cctx.advance_past(drain_offset + token_len);
@@ -955,21 +960,6 @@ impl FmtOp {
             _ => panic!("unknown opcode in fmt data"),
         }
     }
-}
-
-#[inline]
-fn usize_to_u32(value: usize) -> u32 {
-    u32::try_from(value).expect("value must fit in u32")
-}
-
-#[inline]
-fn byte_offset_in(source: &str, ptr: *const u8) -> u32 {
-    let base = source.as_ptr() as usize;
-    let start = ptr as usize;
-    let offset = start
-        .checked_sub(base)
-        .expect("span pointer must be within source");
-    usize_to_u32(offset)
 }
 
 enum GroupNestFrame {
